Rename `RestrictionUtils` to `PermissionUtils`
