+++ src/handlers/interactions/buttons/Manager.ts
@@ -2,7 +2,7 @@ import ClientManager from "../../../Client";
 import Button from "./Button";
 
 import {ButtonInteraction, Collection, TextChannel} from "discord.js";
-import {hasInteractionPermission} from "../../../utils/RestrictionUtils";
+import {hasInteractionPermission} from "../../../utils/PermissionUtils";
 import {InteractionResponseType} from "../../../utils/Types";
 import {sendLog} from "../../../utils/LoggingUtils";
 import {readdir} from "node:fs/promises";
